rename hook, extract reducer and types, tweak var names and action
